ref: Reorganize and document Caching functionality
