+++ crates/symbolicator-service/src/caching/memory.rs
@@ -13,10 +13,10 @@ use symbolic::common::ByteView;
 use tempfile::NamedTempFile;
 use tokio::fs;
 
-pub use super::cache_key::CacheKey;
-use crate::cache::{Cache, CacheEntry, CacheError, ExpirationTime};
-use crate::services::shared_cache::{CacheStoreReason, SharedCacheKey, SharedCacheRef};
+use super::shared_cache::{CacheStoreReason, SharedCacheKey, SharedCacheRef};
 use crate::utils::futures::CallOnDrop;
+
+use super::{Cache, CacheEntry, CacheError, CacheKey, ExpirationTime};
 
 type ComputationChannel<T> = Shared<oneshot::Receiver<CacheEntry<T>>>;
 type ComputationMap<T> = Arc<Mutex<BTreeMap<CacheKey, ComputationChannel<T>>>>;
@@ -522,198 +522,4 @@ async fn persist_tempfile(
         }
     };
     Ok(file)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::atomic::{AtomicIsize, AtomicUsize, Ordering};
-    use std::time::Duration;
-
-    use crate::cache::CacheName;
-    use crate::config::{CacheConfig, CacheConfigs};
-    use crate::test;
-
-    use super::*;
-
-    #[derive(Clone, Default)]
-    struct TestCacheItem {
-        computations: Arc<AtomicUsize>,
-    }
-
-    impl TestCacheItem {
-        fn new() -> Self {
-            Self {
-                computations: Default::default(),
-            }
-        }
-    }
-
-    impl CacheItemRequest for TestCacheItem {
-        type Item = String;
-
-        const VERSIONS: CacheVersions = CacheVersions {
-            current: 1,
-            fallbacks: &[0],
-        };
-
-        fn compute<'a>(&'a self, temp_file: &'a mut NamedTempFile) -> BoxFuture<'a, CacheEntry> {
-            self.computations.fetch_add(1, Ordering::SeqCst);
-
-            Box::pin(async move {
-                tokio::time::sleep(Duration::from_millis(100)).await;
-
-                std::fs::write(temp_file.path(), "some new cached contents")?;
-                Ok(())
-            })
-        }
-
-        fn load(
-            &self,
-            data: ByteView<'static>,
-            _expiration: ExpirationTime,
-        ) -> CacheEntry<Self::Item> {
-            Ok(std::str::from_utf8(data.as_slice()).unwrap().to_owned())
-        }
-    }
-
-    /// This test asserts that the cache is served from outdated cache files, and that a computation
-    /// is being kicked off (and deduplicated) in the background
-    #[tokio::test]
-    async fn test_cache_fallback() {
-        test::setup();
-
-        let cache_dir = test::tempdir().path().join("test");
-        std::fs::create_dir_all(cache_dir.join("global")).unwrap();
-        std::fs::write(
-            cache_dir.join("global/some_cache_key"),
-            "some old cached contents",
-        )
-        .unwrap();
-
-        let cache = Cache::from_config(
-            CacheName::Objects,
-            Some(cache_dir),
-            None,
-            CacheConfig::from(CacheConfigs::default().derived),
-            Arc::new(AtomicIsize::new(1)),
-        )
-        .unwrap();
-        let cacher = Cacher::new(cache, Default::default());
-
-        let request = TestCacheItem::new();
-        let key = CacheKey {
-            cache_key: "global/some_cache_key".into(),
-        };
-
-        let first_result = cacher.compute_memoized(request.clone(), key.clone()).await;
-        assert_eq!(first_result.unwrap().as_str(), "some old cached contents");
-
-        let second_result = cacher.compute_memoized(request.clone(), key.clone()).await;
-        assert_eq!(second_result.unwrap().as_str(), "some old cached contents");
-
-        tokio::time::sleep(Duration::from_millis(200)).await;
-
-        let third_result = cacher.compute_memoized(request.clone(), key).await;
-        assert_eq!(third_result.unwrap().as_str(), "some new cached contents");
-
-        // we only want to have the actual computation be done a single time
-        assert_eq!(request.computations.load(Ordering::SeqCst), 1);
-    }
-
-    /// Makes sure that a `NotFound` result does not fall back to older cache versions.
-    #[tokio::test]
-    async fn test_cache_fallback_notfound() {
-        test::setup();
-
-        let cache_dir = test::tempdir().path().join("test");
-        std::fs::create_dir_all(cache_dir.join("global")).unwrap();
-        std::fs::write(
-            cache_dir.join("global/some_cache_key"),
-            "some old cached contents",
-        )
-        .unwrap();
-        std::fs::create_dir_all(cache_dir.join("1/global")).unwrap();
-        std::fs::write(cache_dir.join("1/global/some_cache_key"), "").unwrap();
-
-        let cache = Cache::from_config(
-            CacheName::Objects,
-            Some(cache_dir),
-            None,
-            CacheConfig::from(CacheConfigs::default().derived),
-            Arc::new(AtomicIsize::new(1)),
-        )
-        .unwrap();
-        let cacher = Cacher::new(cache, Default::default());
-
-        let request = TestCacheItem::new();
-        let key = CacheKey {
-            cache_key: "global/some_cache_key".into(),
-        };
-
-        let first_result = cacher.compute_memoized(request.clone(), key).await;
-        assert_eq!(first_result, Err(CacheError::NotFound));
-
-        // no computation should be done
-        assert_eq!(request.computations.load(Ordering::SeqCst), 0);
-    }
-
-    /// This test asserts that the bounded maximum number of recomputations is not exceeded.
-    #[tokio::test]
-    async fn test_lazy_computation_limit() {
-        test::setup();
-
-        let keys = &["global/1", "global/2", "global/3"];
-
-        let cache_dir = test::tempdir().path().join("test");
-        std::fs::create_dir_all(cache_dir.join("global")).unwrap();
-        for key in keys {
-            let path = cache_dir.join(key);
-            std::fs::write(path, "some old cached contents").unwrap();
-        }
-
-        let cache = Cache::from_config(
-            CacheName::Objects,
-            Some(cache_dir),
-            None,
-            CacheConfig::from(CacheConfigs::default().derived),
-            Arc::new(AtomicIsize::new(1)),
-        )
-        .unwrap();
-        let cacher = Cacher::new(cache, Default::default());
-
-        let request = TestCacheItem::new();
-
-        for key in keys {
-            let request = request.clone();
-            let key = CacheKey {
-                cache_key: String::from(*key),
-            };
-
-            let result = cacher.compute_memoized(request.clone(), key).await;
-            assert_eq!(result.unwrap().as_str(), "some old cached contents");
-        }
-
-        tokio::time::sleep(Duration::from_millis(200)).await;
-
-        // we want the actual computation to be done only one time, as that is the
-        // maximum number of lazy computations.
-        assert_eq!(request.computations.load(Ordering::SeqCst), 1);
-
-        // double check that we actually get outdated contents for two of the requests.
-        let mut num_outdated = 0;
-
-        for key in keys {
-            let request = request.clone();
-            let key = CacheKey {
-                cache_key: String::from(*key),
-            };
-
-            let result = cacher.compute_memoized(request.clone(), key).await;
-            if result.unwrap().as_str() == "some old cached contents" {
-                num_outdated += 1;
-            }
-        }
-
-        assert_eq!(num_outdated, 2);
-    }
 }
