+++ crates/symbolicator-service/src/config.rs
@@ -13,7 +13,7 @@ use tracing::level_filters::LevelFilter;
 
 use symbolicator_sources::SourceConfig;
 
-use crate::cache::SharedCacheConfig;
+use crate::caching::SharedCacheConfig;
 
 /// Controls the log format
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Deserialize)]
