+++ crates/symbolicator-service/src/services/bitcode.rs
@@ -16,14 +16,15 @@ use symbolic::debuginfo::macho::{BcSymbolMap, UuidMapping};
 use symbolicator_sources::{FileType, RemoteFile, SourceConfig};
 use tempfile::NamedTempFile;
 
-use crate::cache::{Cache, CacheEntry, CacheError, ExpirationTime};
-use crate::services::cacher::{CacheItemRequest, CacheKey, Cacher};
+use crate::caching::{
+    Cache, CacheEntry, CacheError, CacheItemRequest, CacheKey, Cacher, ExpirationTime,
+    SharedCacheRef,
+};
 use crate::services::download::DownloadService;
 use crate::types::Scope;
 use crate::utils::futures::{m, measure};
 
 use super::fetch_file;
-use super::shared_cache::SharedCacheRef;
 
 /// Handle to a valid BCSymbolMap.
 ///
